Move strip to its own file
